requirements.txt
Django==3.2.18
django-admin-sortable2==1.0.4
django-tinymce==3.6.1
python-dotenv==0.21.1
Pillow==9.4.0
environs==9.5.0
requests==2.28.2
